requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-environ==0.11.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google_search_results==2.4.2
idna==3.7
PyJWT==2.9.0
requests==2.32.3
serpapi==0.1.5
sqlparse==0.5.1
urllib3==2.2.2
celery==5.1.2
whitenoise==6.7.0
